requirements.txt
Django==4.2.7
channels==4.1.0
daphne==4.1.2
djangorestframework==3.14.0
mysqlclient==2.2.4
